Refactor food and product controllers to improve notification handling and remove unused sanitizeDescription function
