+++ src/controllers/foodController.js
@@ -48,7 +48,7 @@ export const createFood = async (req, res) => {
         to: admin._id,
         title: `New Food created`,
         content: `${req.user.fullName} created Food (${newFood.name}).`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
@@ -69,22 +69,21 @@ export const getFoods = async (req, res) => {
   try {
     const filter = {};
 
-
-    if (req.user.role !== 'superadmin') {
+    if (req.user.role !== "superadmin") {
       if (!req.query.menuId && !req.query.workspaceId) {
-        return res.status(400).json({ 
-          error: "Either menuId or workspaceId is required" 
+        return res.status(400).json({
+          error: "Either menuId or workspaceId is required",
         });
       }
     }
 
-if (active === 'false') filter.isActive = false;
+    if (active === "false") filter.isActive = false;
     if (req.query.menuId) filter.menuId = req.query.menuId;
     if (req.query.workspaceId) filter.workspaceId = req.query.workspaceId;
 
     const foods = await Food.find(filter)
-      .populate('menuId', 'name')
-      .populate('workspaceId', 'name')
+      .populate("menuId", "name")
+      .populate("workspaceId", "name");
 
     res.json({ foods });
   } catch (error) {
@@ -127,7 +126,7 @@ export const updateFood = async (req, res) => {
         to: admin._id,
         title: `Food updated`,
         content: `${req.user.fullName} updated Food (${food.name}).`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
@@ -166,11 +165,14 @@ export const approveFood = async (req, res) => {
       to: food.workspaceId.userId,
       title: `Food Approved`,
       content: `(${food.name}) Food has been Approved by our Admin.`,
-      from: req.user.id
+      from: req.user.id,
     });
     await notification.save();
 
-    sendNotification(food.workspaceId.userId, { ...notification.toJSON(), from: req.user });
+    sendNotification(food.workspaceId.userId, {
+      ...notification.toJSON(),
+      from: req.user,
+    });
 
     res.status(200).json({ message: "Food item approved successfully", food });
   } catch (error) {
@@ -202,11 +204,14 @@ export const unapproveFood = async (req, res) => {
       to: food.workspaceId.userId,
       title: `Food Approved`,
       content: `(${food.name}) Food has been Unapproved by our Admin.`,
-      from: req.user.id
+      from: req.user.id,
     });
     await notification.save();
 
-    sendNotification(food.workspaceId.userId, { ...notification.toJSON(), from: req.user });
+    sendNotification(food.workspaceId.userId, {
+      ...notification.toJSON(),
+      from: req.user,
+    });
 
     res
       .status(200)
@@ -237,11 +242,14 @@ export const deleteFood = async (req, res) => {
         to: food.workspaceId.userId,
         title: `Food Deleted`,
         content: `(${food.name}) Food has been Deleted by ${req.user.fullName}.`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
-      sendNotification(food.workspaceId.userId, { ...notification.toJSON(), from: req.user });
+      sendNotification(food.workspaceId.userId, {
+        ...notification.toJSON(),
+        from: req.user,
+      });
     } else {
       const admin = await User.findOne({ role: "superadmin" });
       if (admin) {
@@ -249,11 +257,14 @@ export const deleteFood = async (req, res) => {
           to: admin._id,
           title: `Food Approved`,
           content: `(${food.name}) Food has been Deleted by ${req.user.fullName}.`,
-          from: req.user.id
+          from: req.user.id,
         });
         await notification.save();
 
-        sendNotification(admin._id, { ...notification.toJSON(), from: req.user });
+        sendNotification(admin._id, {
+          ...notification.toJSON(),
+          from: req.user,
+        });
       }
     }
 
@@ -285,7 +296,7 @@ export const createMenu = async (req, res) => {
         to: admin._id,
         title: `New Menu Created`,
         content: `${req.user.fullName} created Menu (${newMenu.name}).`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
@@ -300,32 +311,8 @@ export const createMenu = async (req, res) => {
   }
 };
 
-
-export const sanitizeDescription = (req, res, next) => {
-  if (req.body.description) {
-    req.body.description = sanitizeHtml(req.body.description, {
-      allowedTags: ['b', 'i', 'u', 'em', 'strong', 'p', 'br', 'ul', 'ol', 'li', 'a'],
-      allowedAttributes: {
-        'a': ['href', 'target', 'rel'],
-      },
-      allowedSchemes: ['http', 'https', 'mailto'],
-      transformTags: {
-        'a': (tagName, attribs) => ({
-          tagName: 'a',
-          attribs: {
-            href: attribs.href,
-            target: '_blank',
-            rel: 'noopener noreferrer'
-          }
-        })
-      }
-    });
-  }
-  next();
-};
-
 export const getMenus = async (req, res) => {
-  const {workspaceId} = req.query;
+  const { workspaceId } = req.query;
   try {
     const menus = await Menu.find();
     res.json({ menus });
@@ -363,7 +350,7 @@ export const updateMenu = async (req, res) => {
         to: admin._id,
         title: `Menu Updated`,
         content: `${req.user.fullName} updated menu (${menu.name}).`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
@@ -397,7 +384,7 @@ export const deleteMenu = async (req, res) => {
         to: admin._id,
         title: `Menu Deleted`,
         content: `(${menu.name}) Menu has been Deleted by ${req.user.fullName}.`,
-        from: req.user.id
+        from: req.user.id,
       });
       await notification.save();
 
@@ -415,7 +402,7 @@ export const activateFood = async (req, res) => {
     const food = await Food.findById(req.params.id);
     food.isActive = true;
     await food.save();
-    res.json({ message: 'Food activated', food });
+    res.json({ message: "Food activated", food });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -426,7 +413,7 @@ export const deactivateFood = async (req, res) => {
     const food = await Food.findById(req.params.id);
     food.isActive = false;
     await food.save();
-    res.json({ message: 'Food deactivated', food });
+    res.json({ message: "Food deactivated", food });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
